[sl-core] Complete support for C's (non-standard) `alloca'.
Since `alloca' is merely play on the stack pointer and the frame
pointer everything was in place except the declaration of the name
`alloca' itself. This changeset provides a header for this
purpose.

* lib/include/calloca.h: New.

* lib/include/sac_helpers.h: Add `safe_free()', which calls
`free()' if and only if it detects its argument is not the result
of `alloca'. This is non-portable (Microgrid only).

git-svn-id: svn+ssh://mac-chris/Library/SVN/CSA/sw/svp/sys/trunk@2751 e97e5017-a994-416e-809d-76780e9f78db