[sl-core] Fix bug in MT-alpha code generation.
When the legacy RA register is used as a temporary, a special
"complete memory load" register access is emitted before the point
a thread terminates, to ensure pending memory loads are flushed
before the thread resources are used. This is a "required
optimization" (to avoid pipeline stalls in the next thread that
reuses the register). Unfortunately it is possible that the
control flow reaches that point without using the RA register (it
could be used only in some branches in the function body). The
solution implemented is to conservatively initialize RA always
when it may be used as a temporary.

* tools/bin/spp (pppalpha.makeprotectspecial): New transform.
(pppalpha.filter): Use it.

git-svn-id: svn+ssh://mac-chris/Library/SVN/CSA/sw/svp/sys/trunk@2950 e97e5017-a994-416e-809d-76780e9f78db