[sl-core] Re-optimize accesses to shareds in `ppp' target.
The previous changeset was required to prevent incorrect code, but
would also require a temporary register in all cases, including
the simple `sl_setp(x, sl_getp(x) + 1)' which should instead be
optimized to a single instruction. This changeset reintroduces
this optimization opportunity.

* tools/lib/t-pppalpha/include/sl_support.h: Remove superfluous
clobber of the incoming shared just before a write to the
outgoing. (but still clobber it afterwards)

git-svn-id: svn+ssh://mac-chris/Library/SVN/CSA/sw/svp/sys/trunk@2987 e97e5017-a994-416e-809d-76780e9f78db