[sl-core] New MT-Alpha optimization.
The following was observed during Apple-CORE benchmarking: in many
"small" threads the operation that use the result of a memory load
directly follow the load. They will arrive at the pipeline read
stage without their dependency in most case, and then be
rescheduled. This wastes one execution cycle per thread. It can be
avoided in most cases by forcing a thread switch immediately
before such instructions, so that other threads get a chance to
execute and the dependency be satisfied by the time the
instruction goes around the active thread queue.

* tools/bin/spp (pppalpha.addswchll): Also add `swch' *before*
operations that use the result of a long latency operation (not
only after).

git-svn-id: svn+ssh://mac-chris/Library/SVN/CSA/sw/svp/sys/trunk@3281 e97e5017-a994-416e-809d-76780e9f78db