[sl-core] Support `sl_create' inside plain C functions.
The MT-Alpha target now support the construct; this changeset
extends the functionality to the sequential C target.

* tools/lib/t-seqc/slimpl.m4: Deprecate `sl_kill'. Implement new
`sl_break' semantics (not yet official). This removes the need for
a linked list of family contexts, so a new child can appear
disconnected. This allows `sl_create' to appear inside plain C
functions.

* tools/lib/t-seqc/include/sl_support.h: Remove the obsolete field.

git-svn-id: svn+ssh://mac-chris/Library/SVN/CSA/sw/svp/sys/trunk@3666 e97e5017-a994-416e-809d-76780e9f78db