[sl-base] Modify the stock GCC to better support C99's VLA's.
The stock GNU compiler needs to juggle around the stack pointer
when using large frames. When restoring the frame pointer on exit,
it needs a temporary register. However at this point it does not
honor `-ffixed', so a "forbidden" register could get used. This
changeset prevents this by using the AT register instead.

* deploy/patches/gcc.patch: Extend changes to `alpha.c'.

git-svn-id: svn+ssh://mac-chris/Library/SVN/CSA/sw/svp/sys/trunk@3748 e97e5017-a994-416e-809d-76780e9f78db