[sl-core] Fix bug in `spr'.
A large inefficiency would cause exponential space complexity
during transformation in the depth of C scope nesting. This would
cause exceedingly large compilation times and/or stack blowouts
for large programs.

* tools/lib/slc/visitors.py (ScopedVisitor): Remember the current
stack of C scopes.

* tools/lib/slc/lower/split.py (SplitCreates.visit_lowcreate):
Also memoize the stack of current scopes, so that they are not
duplicated when duplicating the `Create' objects.

git-svn-id: svn+ssh://mac-chris/Library/SVN/CSA/sw/svp/sys/trunk@3906 e97e5017-a994-416e-809d-76780e9f78db